systemd runner: only wait systemd service to start up for 60s
In an incident a previous build failed with timeout and got killed, but
didn't succeed to kill some of those processes because of EPERM. A later
build service doesn't start without any traces. It seems that remaining
processes caused systemd to think the service was still running and
didn't respond to a new start.

We just make sure we don't wait forever. I'll investigate what to do
with those briefly remaining processes later.